requirements.txt
flask>=3.0
python-docx>=1.1
python-pptx>=1.0
gunicorn>=22.0
